Proteger rutas y eliminar landing page antiguo
